requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
colorama
ctransformers
filelock
frozenlist
fsspec
greenlet
h11
httpcore
httpx
huggingface-hub
idna
joblib
jsonpatch
jsonpointer
langchain
langchain-core
langchain-text-splitters
langchain-huggingface
langsmith
MarkupSafe
mpmath
multidict
networkx
numpy
orjson
packaging
pillow
pinecone
propcache
py-cpuinfo
pydantic
pydantic_core
pypdf
PyYAML
regex
requests
requests-toolbelt
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sniffio
SQLAlchemy
streamlit
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
yarl
zstandard
python-dotenv
